+++ scifit/tests/test_solvers_scientific.py
@@ -406,6 +406,7 @@ class GaussianPeakWithBaselineRegressionNoiseL2(
 
 
 class EMGPeakRegression(GenericLinearRegression):
+    """Heavy duty model"""
     factory = scientific.EMGPeakFitSolver
     configuration = {"p0": (500., 5., 20., 10.)} #, "bounds": [(1e-2, 1e-2, 1e-2, 1e-2), (1e3, 1e3, 1e3, 1e3)]}  # 20 min run
     parameters = np.array([450.3, 1.23, 15.7, 3.42])
@@ -417,20 +418,20 @@ class EMGPeakRegression(GenericLinearRegression):
     def test_fit_from_synthetic_dataset(self):
         pass
 
-
-class EMGPeakRegressionNoiseL0(
-    EMGPeakRegression, TestCase
-):
-    sigma = 1e-6
-
-
-class EMGPeakRegressionNoiseL1(
-    EMGPeakRegression, TestCase
-):
-    sigma = 2.5e-2
-
-
-class EMGPeakRegressionNoiseL2(
-    EMGPeakRegression, TestCase
-):
-    sigma = 1e-1
+#
+# class EMGPeakRegressionNoiseL0(
+#     EMGPeakRegression, TestCase
+# ):
+#     sigma = 1e-6
+#
+#
+# class EMGPeakRegressionNoiseL1(
+#     EMGPeakRegression, TestCase
+# ):
+#     sigma = 2.5e-2
+#
+#
+# class EMGPeakRegressionNoiseL2(
+#     EMGPeakRegression, TestCase
+# ):
+#     sigma = 1e-1